feat: findAllByBusinessId 메서드를 추가하여 특정 사업체의 상품 전체 조회시에 같은 바코드, 같은 층, 같은 로케이션 명,같은 유통기한을 가지는 경우 상품의 총 개수를 한 행에 출력하는 기능 추가

+++ wms_backend/src/main/java/com/a508/wms/productdetail/domain/ProductDetail.java
@@ -35,19 +35,19 @@ public class ProductDetail extends BaseTimeEntity {
     @Column(nullable = false)
     private String name;
     @Column
-    private Long size;
+    private Integer size;
     @Column
-    private Long unit;
+    private Integer unit;
     @Column
-    private int originalPrice;
+    private Integer originalPrice;
     @Column
-    private int sellingPrice;
+    private Integer sellingPrice;
     @Enumerated(EnumType.STRING)
     private StatusEnum statusEnum = StatusEnum.ACTIVE;
 
     @Builder
     public ProductDetail(Business business, ProductStorageTypeEnum productStorageType, Long barcode,
-                         String name, Long size, Long unit, int originalPrice, int sellingPrice) {
+                         String name, int size, int unit, int originalPrice, int sellingPrice) {
         this.business = business;
         this.productStorageType = productStorageType;
         this.barcode = barcode;
@@ -68,7 +68,7 @@ public class ProductDetail extends BaseTimeEntity {
     }
 
     public void updateData(ProductStorageTypeEnum productStorageType
-            , Long barcode, String name, Long size, Long unit
+            , Long barcode, String name, int size, int unit
             , int originalPrice, int sellingPrice) {
         this.productStorageType = productStorageType;
         this.barcode = barcode;
